#4 refactor: 코드개선 및 공통요소 분리
