+++ client/src/components/compete/Common/Banner.js
