+++ client/src/components/compete/Common/Card.js
