+++ client/src/components/compete/CompDetail.js
@@ -1,11 +1,22 @@
 import { useLocation } from "react-router-dom";
+import styled from "styled-components";
+
+const ButtonWrapper = styled.div`
+  width: 375px;
+  height: 44px;
+  background-color: yellow;
+`;
+
+//배팅 카드
+
+//
 
 function CompDetail(props) {
   const { state } = useLocation();
 
-  //const id = props ? props.match.params.id : "none";
   console.log(state);
-  return <div>세부 페이지입니다</div>;
+
+  return <ButtonWrapper>뒤로가기</ButtonWrapper>;
 }
 
 export default CompDetail;
