+++ client/src/components/compete/Compete.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Header from "./CompHeader";
+import Header from "./Common/KeyBellHeader";
 import Body from "./CompBody";
 
 function Compete() {
